feat: session lazy join room and disconnect
