+++ index.html
@@ -4,21 +4,12 @@
         <meta charset="UTF-8" />
         <link rel="icon" type="image/svg+xml" href="/vite.svg" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Vite + TS</title>
+        <title>Samples</title>
     </head>
     <body>
         <div id="app"></div>
-        <button id="connect">Connect</button>
-        <audio controls autoplay id="audio-main"></audio>
-        <video
-            controls
-            muted
-            autoplay
-            width="500"
-            height="500"
-            style="background-color: gray"
-            id="video-main"
-        ></video>
+        <div><a href="/samples/echo-fast/">Echo join room fast</a></div>
+        <div><a href="/samples/echo-lazy/">Echo join room lazy</a></div>
         <script type="module" src="/src/main.ts"></script>
     </body>
 </html>
