+++ src/samples/echo_lazy.ts
@@ -1,0 +1,76 @@
+import {
+  RoomTrackStarted,
+  RoomTrackStopped,
+  Session,
+  SessionEvent,
+} from "../../lib/main";
+
+async function connect(_e: any) {
+  const session = new Session("http://localhost:3000", { token: "demo-token" });
+  const stream = await navigator.mediaDevices.getUserMedia({
+    audio: true,
+    video: true,
+  });
+  let audio_send_track = session.sender(
+    "audio_main",
+    stream.getAudioTracks()[0],
+    100,
+  );
+  let video_send_track = session.sender(
+    "video_main",
+    stream.getVideoTracks()[0],
+    100,
+  );
+  console.log(audio_send_track, video_send_track);
+  let audio_recv_track = session.receiver("audio", 100);
+  let video_recv_track = session.receiver("video", 100);
+  (document.getElementById("audio-main")! as HTMLAudioElement).srcObject =
+    audio_recv_track.stream;
+  (document.getElementById("video-main")! as HTMLVideoElement).srcObject =
+    video_recv_track.stream;
+
+  session.on(SessionEvent.ROOM_TRACK_STARTED, (track: RoomTrackStarted) => {
+    console.log("Track started", track);
+    if (track.track == "audio_main") {
+      audio_recv_track.attach(track).then(console.log).catch(console.error);
+    } else {
+      video_recv_track.attach(track).then(console.log).catch(console.error);
+    }
+  });
+
+  session.on(SessionEvent.ROOM_TRACK_STOPPED, (track: RoomTrackStopped) => {
+    console.log("Track stopped", track);
+  });
+
+  document.getElementById("join")!.onclick = () => {
+    session
+      .join(
+        {
+          room: "demo",
+          peer: "web-1",
+          publish: { peer: true, tracks: true },
+          subscribe: { peers: true, tracks: true },
+        },
+        "token",
+      )
+      .then(console.log)
+      .catch(console.error);
+  };
+
+  document.getElementById("leave")!.onclick = () => {
+    session.leave().then(console.log).catch(console.error);
+  };
+  document.getElementById("disconnect")!.onclick = () => {
+    stream.getTracks().map((t) => {
+      t.stop();
+    });
+    (document.getElementById("audio-main")! as HTMLAudioElement).srcObject =
+      null;
+    (document.getElementById("video-main")! as HTMLVideoElement).srcObject =
+      null;
+    session.disconnect();
+  };
+  await session.connect();
+}
+
+document.getElementById("connect")!.onclick = connect;
