+++ vite-build.config.ts
@@ -1,0 +1,31 @@
+// vite.config.js
+import { resolve } from "path";
+import { defineConfig } from "vite";
+import dts from "vite-plugin-dts";
+
+export default defineConfig({
+  plugins: [dts({ include: ["lib"] })],
+  define: {
+    SDK_VERSION: JSON.stringify(process.env.npm_package_version),
+  },
+  build: {
+    copyPublicDir: false,
+    lib: {
+      // Could also be a dictionary or array of multiple entry points
+      entry: resolve(__dirname, "lib/main.ts"),
+      name: "Atm0sMediaSdk",
+      // the proper extensions will be added
+      fileName: "atm0s-media-sdk",
+    },
+    rollupOptions: {
+      // make sure to externalize deps that shouldn't be bundled
+      // into your library
+      external: [],
+      output: {
+        // Provide global variables to use in the UMD build
+        // for externalized deps
+        globals: {},
+      },
+    },
+  },
+});
