+++ vite.config.ts
@@ -9,22 +9,10 @@ export default defineConfig({
     SDK_VERSION: JSON.stringify(process.env.npm_package_version),
   },
   build: {
-    copyPublicDir: false,
-    lib: {
-      // Could also be a dictionary or array of multiple entry points
-      entry: resolve(__dirname, "lib/main.ts"),
-      name: "Atm0sMediaSdk",
-      // the proper extensions will be added
-      fileName: "atm0s-media-sdk",
-    },
     rollupOptions: {
-      // make sure to externalize deps that shouldn't be bundled
-      // into your library
-      external: [],
-      output: {
-        // Provide global variables to use in the UMD build
-        // for externalized deps
-        globals: {},
+      input: {
+        main: resolve(__dirname, "index.html"),
+        echo: resolve(__dirname, "samples/echo/index.html"),
       },
     },
   },
